requirements.txt
Pillow==9.1.0
selenium==4.1.3
validators==0.18.2
